[firejail] Implement template support. OMP#JOLLA-88 JB#54034
Add two patches which change profile loading to be delayed until all
arguments are parsed. This is due to the template key replacement, since
templates are to be given as arguments as well. Each line in the profile
line is checked and if one template is found they are replaced with the
corresponding value. If no changes are required the profile line is
untouched.

Signed-off-by: Jussi Laakkonen <jussi.laakkonen@jolla.com>