[firejail] Retain symlink chains. Fixes JB#55981
When firejail is constructing e.g. private-etc directory content,
symlink targets are fully resolved during duplication. This breaks all
symlink chains and can be problematic if such chains are intentional.

For example Sailfish OS timezone setup relies on chain like
/etc/localtime -> /var/lib/timed/localtime -> /usr/share/zoneinfo/xxx.
Where the intermediate part is in a time daemon writable directory,
and having firejail squeeze it into /etc/localtime ->
/usr/share/zoneinfo/xxx effectively freezes timezone of sandboxed
applications to whatever it was when the application was started.

Similar problems occur with symlinks that go though dynamic/virtual
entries such as /proc/self.

Retain symlink chains by duplicating symlinks as-is instead of
resolving the final target.

Remove the now unnecessary special case code that was used to rectify
/proc/PID/xxx type symlinks back into /proc/self/xxx form.

Signed-off-by: Simo Piiroinen <simo.piiroinen@jolla.com>